fix agentic chat
