+++ backend/chatservice/service/service.go
@@ -24,7 +24,6 @@ import (
 	"sortedstartup/chatservice/queue"
 	"sortedstartup/chatservice/rag"
 	settings "sortedstartup/chatservice/settings"
-	"sortedstartup/chatservice/sortedagents"
 	"sortedstartup/chatservice/store"
 	"sortedstartup/chatservice/types"
 
@@ -169,9 +168,13 @@ func (s *ChatService) Chat(ctx context.Context, userID string, req *pb.ChatReque
 	}
 
 	provider := req.GetProvider()
-	var providerSettings *pb.ProviderSettings
+	providerSettings, err := s.settingsManager.GetProviderSetting(provider)
+	if err != nil {
+		slog.Error("service:Chat", "error", "failed to get provider settings", "error", err, "provider", provider)
+		return fmt.Errorf("failed to get provider settings")
+	}
+
 	if provider != LOCAL_PROVIDER {
-		providerSettings, err = s.settingsManager.GetProviderSetting(provider)
 		if providerSettings == nil || providerSettings.ApiKey == "" {
 
 			slog.Error("service:Chat", "error", "API key not set for provider", "provider", provider)
@@ -185,11 +188,6 @@ func (s *ChatService) Chat(ctx context.Context, userID string, req *pb.ChatReque
 			})
 			return fmt.Errorf("API key not configured for provider %s, Please configure it in the models page", provider)
 		}
-		if err != nil {
-			slog.Error("service:Chat", "error", "failed to get provider settings", "error", err, "provider", provider)
-			return fmt.Errorf("failed to get provider settings")
-		}
-
 	}
 
 	chatId := req.ChatId
@@ -457,7 +455,7 @@ func (s *ChatService) Chat(ctx context.Context, userID string, req *pb.ChatReque
 		currentUserPrompt = enhancedPrompt
 	}
 
-	if s.shouldUseAgenticChat(provider, hasImages, history, providerSettings) {
+	if s.shouldUseAgenticChat(capabilities, hasImages, history, providerSettings) {
 		return s.runAgenticChat(
 			ctx,
 			chatId,
@@ -663,314 +661,6 @@ func (s *ChatService) Chat(ctx context.Context, userID string, req *pb.ChatReque
 	}
 
 	return nil
-}
-
-const chatAgentInstructions = `You are the regular SortedChat assistant.
-
-Answer directly when the request can be handled from the conversation alone.
-Use the web_search tool only when fresh or external information is actually needed.
-When you use web search, ground the final answer in the search results and include relevant source URLs in the response.`
-
-func (s *ChatService) shouldUseAgenticChat(provider string, hasImages bool, history []dao.ChatMessageRow, providerSettings *pb.ProviderSettings) bool {
-	slog.Debug("service:Chat", "message", "checking if agentic chat should be used", "provider", provider)
-	if hasImages || provider == LOCAL_PROVIDER || providerSettings == nil || strings.TrimSpace(providerSettings.ApiKey) == "" {
-		return false
-	}
-
-	if strings.TrimSpace(os.Getenv("BRAVE_SEARCH_API_KEY")) == "" {
-		return false
-	}
-
-	// TODO: Replace this placeholder gate with model capability metadata when available.
-	modelSupportsAgenticActions := true
-	// modelSupportsAgenticActions := false
-
-	if !modelSupportsAgenticActions {
-		return false
-	}
-
-	for _, msg := range history {
-		if _, ok := extractTextOnlyChatMessage(msg); !ok {
-			return false
-		}
-	}
-
-	return true
-}
-
-func extractTextOnlyChatMessage(msg dao.ChatMessageRow) (string, bool) {
-	slog.Debug("service:Chat", "message", "extracting text from chat message", "messageId", msg)
-	if msg.ContentImage != "" {
-		return "", false
-	}
-
-	if strings.TrimSpace(msg.Content) == "" {
-		return "", true
-	}
-
-	if !(strings.HasPrefix(msg.Content, "[") && strings.HasSuffix(msg.Content, "]")) {
-		return msg.Content, true
-	}
-
-	var contents []*pb.MessageContent
-	if err := json.Unmarshal([]byte(msg.Content), &contents); err != nil {
-		return "", false
-	}
-
-	var textParts []string
-	for _, content := range contents {
-		if content.Type != "text" {
-			return "", false
-		}
-		if strings.TrimSpace(content.Text) != "" {
-			textParts = append(textParts, content.Text)
-		}
-	}
-
-	return strings.TrimSpace(strings.Join(textParts, "\n")), true
-}
-
-func (s *ChatService) buildAgenticSession(chatID string, history []dao.ChatMessageRow) (sortedagents.Session, error) {
-	slog.Debug("service:Chat", "message", "building agentic session from chat history", "chatId", chatID)
-	messages := []sortedagents.Message{
-		{Role: "system", Content: chatAgentInstructions},
-	}
-
-	for _, msg := range history {
-		text, ok := extractTextOnlyChatMessage(msg)
-		if !ok {
-			return nil, fmt.Errorf("chat history contains non-text content")
-		}
-
-		messages = append(messages, sortedagents.Message{
-			Role:    msg.Role,
-			Content: text,
-		})
-	}
-
-	return sortedagents.NewSessionFromMessages(chatID, messages), nil
-}
-
-func (s *ChatService) runAgenticChat(
-	ctx context.Context,
-	chatID string,
-	userID string,
-	projectID string,
-	model string,
-	prompt string,
-	history []dao.ChatMessageRow,
-	ragChunks []rag.Result,
-	ragEnabled bool,
-	providerSettings *pb.ProviderSettings,
-	stream func(*pb.ChatResponse) error,
-) error {
-	slog.Debug("service:Chat", "message", "running agentic chat", "chatId", chatID, "userID", userID, "projectID", projectID)
-	session, err := s.buildAgenticSession(chatID, history)
-	if err != nil {
-		slog.Error("service:Chat", "message", "failed to build agentic session", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-		return fmt.Errorf("failed to prepare chat context")
-	}
-
-	agent := sortedagents.NewAgent(
-		"sortedchat-chat-agent",
-		chatAgentInstructions,
-		model,
-		[]sortedagents.Tool{sortedagents.NewBraveSearchTool()},
-	)
-	runner := sortedagents.NewRunnerWithLLM(sortedagents.NewOpenAILLMWithConfig(
-		providerSettings.ApiKey,
-		providerSettings.ApiUrl,
-		model,
-	))
-
-	if err := stream(&pb.ChatResponse{
-		Response: &pb.ChatResponse_Progress{
-			Progress: &pb.ChatProgress{
-				State:   pb.ChatProgress_SENDING_REQUEST_TO_LLM,
-				Message: "Sending request to LLM",
-			},
-		},
-	}); err != nil {
-		return fmt.Errorf("error while processing request, please try again")
-	}
-
-	events := runner.RunStream(ctx, agent, prompt, 4, session)
-
-	if err := stream(&pb.ChatResponse{
-		Response: &pb.ChatResponse_Progress{
-			Progress: &pb.ChatProgress{
-				State:   pb.ChatProgress_REQUEST_SENT_TO_LLM,
-				Message: "Request sent",
-			},
-		},
-	}); err != nil {
-		return fmt.Errorf("error while processing request, please try again")
-	}
-
-	var fullResponse strings.Builder
-	var inputTokens, outputTokens, cachedTokens int
-	firstToken := true
-	completed := false
-
-	savePartialResponse := func() {
-		assistantText := fullResponse.String()
-		if assistantText == "" {
-			return
-		}
-
-		var referencesJSON string
-		if len(ragChunks) > 0 {
-			ragDocs := s.createRAGDocumentJSONFromChunks(ragChunks)
-			referencesBytes, err := json.Marshal(ragDocs)
-			if err != nil {
-				slog.Error("service:Chat", "message", "failed to marshal RAG document references for partial agentic response", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-			} else {
-				referencesJSON = string(referencesBytes)
-			}
-		}
-
-		nonCachedInputTokens := inputTokens - cachedTokens
-		if nonCachedInputTokens < 0 {
-			nonCachedInputTokens = 0
-		}
-
-		if _, err := s.dao.AddChatMessageWithTokens(userID, chatID, "assistant", assistantText, "", model, nonCachedInputTokens, outputTokens, cachedTokens, referencesJSON, ragEnabled); err != nil {
-			slog.Error("service:Chat", "message", "failed to save partial agentic assistant message", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-		}
-	}
-
-	for event := range events {
-		switch e := event.(type) {
-		case *sortedagents.TextChunkEvent:
-			if firstToken {
-				if err := stream(&pb.ChatResponse{
-					Response: &pb.ChatResponse_Progress{
-						Progress: &pb.ChatProgress{
-							State:   pb.ChatProgress_FIRST_TOKEN_RECEIVED,
-							Message: "First token received",
-						},
-					},
-				}); err != nil {
-					savePartialResponse()
-					return fmt.Errorf("error while processing request, please try again")
-				}
-				firstToken = false
-			}
-
-			fullResponse.WriteString(e.Chunk)
-			if err := stream(&pb.ChatResponse{
-				Response: &pb.ChatResponse_Text{Text: e.Chunk},
-			}); err != nil {
-				savePartialResponse()
-				return fmt.Errorf("error while processing request, please try again")
-			}
-
-		case *sortedagents.ToolCallStartEvent:
-			if e.ToolName != "web_search" {
-				continue
-			}
-			if err := stream(&pb.ChatResponse{
-				Response: &pb.ChatResponse_Progress{
-					Progress: &pb.ChatProgress{
-						State:   pb.ChatProgress_REQUEST_SENT_TO_LLM,
-						Message: "Searching the web",
-					},
-				},
-			}); err != nil {
-				savePartialResponse()
-				return fmt.Errorf("error while processing request, please try again")
-			}
-
-		case *sortedagents.UsageEvent:
-			inputTokens = e.InputTokens
-			outputTokens = e.OutputTokens
-			cachedTokens = e.CachedTokens
-
-		case *sortedagents.CompleteEvent:
-			completed = true
-
-		case *sortedagents.ErrorEvent:
-			if ctx.Err() != nil {
-				slog.Info("service:Chat", "message", "agentic streaming cancelled by client, saving partial response", "error", e.Error, "chatId", chatID, "userID", userID, "projectID", projectID)
-			} else {
-				slog.Error("service:Chat", "message", "agentic chat failed", "error", e.Error, "chatId", chatID, "userID", userID, "projectID", projectID)
-			}
-			savePartialResponse()
-			return fmt.Errorf("error while processing request, please try again")
-		}
-	}
-
-	if !completed && ctx.Err() != nil {
-		savePartialResponse()
-		return fmt.Errorf("error while processing request, please try again")
-	}
-
-	if err := stream(&pb.ChatResponse{
-		Response: &pb.ChatResponse_Progress{
-			Progress: &pb.ChatProgress{
-				State:   pb.ChatProgress_TOKENS_STOPPED,
-				Message: "Response finished",
-			},
-		},
-	}); err != nil {
-		return fmt.Errorf("error while processing request, please try again")
-	}
-
-	assistantText := fullResponse.String()
-	if assistantText != "" {
-		var referencesJSON string
-		if len(ragChunks) > 0 {
-			ragDocs := s.createRAGDocumentJSONFromChunks(ragChunks)
-			referencesBytes, err := json.Marshal(ragDocs)
-			if err != nil {
-				slog.Error("service:Chat", "message", "failed to marshal RAG document references for final agentic response", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-			} else {
-				referencesJSON = string(referencesBytes)
-			}
-		}
-
-		nonCachedInputTokens := inputTokens - cachedTokens
-		if nonCachedInputTokens < 0 {
-			nonCachedInputTokens = 0
-		}
-
-		daoSummary, err := s.dao.AddChatMessageWithTokens(userID, chatID, "assistant", assistantText, "", model, nonCachedInputTokens, outputTokens, cachedTokens, referencesJSON, ragEnabled)
-		if err != nil {
-			slog.Error("service:Chat", "message", "failed to insert agentic assistant message", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-		} else {
-			pbSummary := &pb.ResponseSummary{
-				MessageId:    daoSummary.MessageId,
-				Model:        model,
-				InputTokens:  int32(daoSummary.InputTokenCount),
-				OutputTokens: int32(daoSummary.OutputTokenCount),
-				CachedTokens: int32(daoSummary.CachedTokenCount),
-				Cost:         float32(daoSummary.Cost),
-			}
-			if err := stream(&pb.ChatResponse{
-				Response: &pb.ChatResponse_Summary{Summary: pbSummary},
-			}); err != nil {
-				return fmt.Errorf("failed to send message summary, please try again")
-			}
-		}
-	}
-
-	chatInfo, err := s.dao.GetChatMetadata(userID, chatID)
-	if err != nil {
-		slog.Error("service:Chat", "message", "failed to get chat metadata after agentic response", "error", err, "chatId", chatID, "userID", userID, "projectID", projectID)
-		return fmt.Errorf("error while processing request, please try again")
-	}
-
-	return stream(&pb.ChatResponse{
-		Response: &pb.ChatResponse_ChatMetadata{
-			ChatMetadata: &pb.ChatInfo{
-				ChatId:           chatID,
-				Cost:             float32(chatInfo.Cost),
-				InputTokenCount:  int32(chatInfo.InputTokenCount),
-				OutputTokenCount: int32(chatInfo.OutputTokenCount),
-				CachedTokenCount: int32(chatInfo.CachedTokenCount),
-			},
-		},
-	})
 }
 
 const (
